成功修改Y为稀疏矩阵，但是有cf(8,3)->cf(8,-3)的bug

+++ Equipment.h
@@ -2,7 +2,7 @@
 #include <complex>
 #include <cstddef>
 
-using NodeType = size_t;
+using NodeType = int;
 using DeviceArgType = float;
 using cf = std::complex<DeviceArgType>;
 using socketType = std::pair<NodeType, NodeType>;
